// types/nexus-typegen.ts
export interface NexusGenInputs {
  LoginInput: {
    // input type
    email: string // String!
    password: string // String!
  }
  SignupInput: {
    // input type
    email: string // String!
    name: string // String!
    password: string // String!
  }
  UserUpdateInput: {
    // input type
    email: NexusGenScalars["EmailAddress"] // EmailAddress!
    name: string // String!
    password: string // String!
  }
}

export interface NexusGenEnums {
  Role: "ADMIN" | "USER"
}

export interface NexusGenScalars {
  String: string
  Int: number
  Float: number
  Boolean: boolean
  ID: string
  DateTime: any
  EmailAddress: any
}

export interface NexusGenObjects {
  AuthPayload: {
    // root type
    token: string // String!
    user: NexusGenRootTypes["User"] // User!
  }
  Mutation: {}
  Query: {}
  User: {
    // root type
    email: NexusGenScalars["EmailAddress"] // EmailAddress!
    id: string // ID!
    name: string // String!
    role: NexusGenEnums["Role"] // Role!
  }
}

export interface NexusGenInterfaces {}

export interface NexusGenUnions {}

export type NexusGenRootTypes = NexusGenObjects
// ...
export interface NexusGenArgTypes {
  Mutation: {
    deleteUser: {
      // args
      id: string // ID!
    }
    login: {
      // args
      data: NexusGenInputs["LoginInput"] // LoginInput!
    }
    signup: {
      // args
      data: NexusGenInputs["SignupInput"] // SignupInput!
    }
    updateUser: {
      // args
      data: NexusGenInputs["UserUpdateInput"] // UserUpdateInput!
      id: string // ID!
    }
  }
  Query: {
    getUser: {
      // args
      id: string // ID!
    }
  }
}